requirements.txt
thonny
websocket_client